First Local Vision Math Commit - Fails Under Certain Current Game Conditions
Does not Work if The Camera is ABOVE the target - New Math to Come

+++ vision_app/app/src/main/jni/image_processor.cpp
@@ -158,7 +158,7 @@ std::vector<TargetInfo> processImpl(int w, int h, int texOut, DisplayMode mode,
         }
     }
 
-    LOGD("Total Number of Targets: %d", accepted_targets.size());
+    LOGD("Total Number of Targets: %d", (int) accepted_targets.size());
 
     //Takes the top so many blocks in size
     const int kMaxTargets = 6;
